hrms-backend/requirements.txt
Django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10; python_version < "3.13"
psycopg[binary]==3.2.13; python_version >= "3.13"
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2
